[refactor] 패키지명 수정
패키지를 application-domain-persistence-presentaion 구조로 설정하였습니다.
List<Category> 대신, CategoryInfo를 구현해서 dto로 적용시켰습니다.

관련 이슈: #7
