+++ src/main/java/team7/inplace/place/application/dto/CategoryInfo.java
@@ -1,0 +1,5 @@
+package team7.inplace.place.application.dto;
+
+public record CategoryInfo(String name) {
+
+}
