--- src/main/java/team7/inplace/place/domain/CategoryListResponse.java
@@ -1,7 +1,0 @@
-package team7.inplace.place.domain;
-
-import java.util.List;
-
-public record CategoryListResponse(List<Category> categories) {
-
-}
