+++ src/main/java/team7/inplace/place/presentation/PlaceController.java
@@ -1,4 +1,4 @@
-package team7.inplace.place.controller;
+package team7.inplace.place.presentation;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -7,9 +7,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import team7.inplace.place.domain.Category;
-import team7.inplace.place.domain.CategoryListResponse;
-import team7.inplace.place.service.PlaceService;
+import team7.inplace.place.application.PlaceService;
+import team7.inplace.place.application.dto.CategoryInfo;
+import team7.inplace.place.presentation.dto.CategoriesResponse;
 
 @RestController
 @RequiredArgsConstructor
@@ -19,9 +19,9 @@ public class PlaceController {
     private final PlaceService placeService;
 
     @GetMapping("/categories")
-    public ResponseEntity<CategoryListResponse> getCategories() {
-        List<Category> categories = placeService.getCategories();
-        CategoryListResponse response = new CategoryListResponse(categories);
+    public ResponseEntity<CategoriesResponse> getCategories() {
+        List<CategoryInfo> categories = placeService.getCategories();
+        CategoriesResponse response = new CategoriesResponse(categories);
 
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
