+++ src/main/java/team7/inplace/place/presentation/dto/CategoriesResponse.java
@@ -1,0 +1,8 @@
+package team7.inplace.place.presentation.dto;
+
+import java.util.List;
+import team7.inplace.place.application.dto.CategoryInfo;
+
+public record CategoriesResponse(List<CategoryInfo> categories) {
+
+}
